WIP: RPC refactor
