+++ packages/syft-rpc/syft_rpc/__init__.py
@@ -1,6 +1,3 @@
-from .json import JSONModel  # noqa: F401
-from .rpc import Future  # noqa: F401
-from .rpc import Request  # noqa: F401
-from .rpc import SyftBoxURL  # noqa: F401
+from .rpc import send  # noqa: F401
 
 __version__ = "0.1.0"
